feat: eneble parallel testing on a custom collection
