+++ Meziantou.Xunit.ParallelTestFramework.Tests/ParallelCollectionAttributeTests.cs
@@ -1,0 +1,26 @@
+using Xunit;
+
+namespace Meziantou.Xunit.ParallelTestFramework.Tests;
+
+[Collection("Parallel")]
+public class ParallelCollectionAttributeTests : IClassFixture<ConcurrencyFixture>
+{
+    private readonly ConcurrencyFixture fixture;
+
+    public ParallelCollectionAttributeTests(ConcurrencyFixture fixture)
+    {
+        this.fixture = fixture;
+    }
+
+    [Fact]
+    public async Task Fact1()
+    {
+        Assert.Equal(2, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
+    }
+
+    [Fact]
+    public async Task Fact2()
+    {
+        Assert.Equal(2, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
+    }
+}
