+++ Meziantou.Xunit.ParallelTestFramework/ParallelTestClassRunner.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
-using Xunit;
 using Xunit.Abstractions;
 using Xunit.Sdk;
 
@@ -16,8 +15,7 @@ public class ParallelTestClassRunner : XunitTestClassRunner
     // https://github.com/xunit/xunit/blob/2.4.2/src/xunit.execution/Sdk/Frameworks/Runners/TestClassRunner.cs#L194-L219
     protected override async Task<RunSummary> RunTestMethodsAsync()
     {
-        var disableParallelization = TestClass.Class.GetCustomAttributes(typeof(DisableParallelizationAttribute)).Any()
-            || TestClass.Class.GetCustomAttributes(typeof(CollectionAttribute)).Any();
+        var disableParallelization = TestClass.Class.GetCustomAttributes(typeof(DisableParallelizationAttribute)).Any();
 
         if (disableParallelization)
             return await base.RunTestMethodsAsync().ConfigureAwait(false);
